feat(dashboard): a suspended app says so instead of failing a browse
Its files are read inside the microVM that has them mounted, so there is nothing to ask.
